Implement 32-bit hash combine function from Boost.

+++ CesiumUtility/include/CesiumUtility/Hash.h
@@ -18,7 +18,7 @@ struct Hash {
    * @param second The second hash value.
    * @return A new hash value which is a combination of the two.
    */
-  static size_t combine(uint64_t first, uint64_t second);
+  static size_t combine(size_t first, size_t second);
 };
 
 } // namespace CesiumUtility
